requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==0.27.0
pandas==1.4.0
numpy==1.21.0
scikit-learn==0.24.2
requests==2.26.0
flask-cors==4.0.0
gunicorn==20.1.0
